+++ ExchangeApp.BL/Models/Currency/CurrencyDetailModel.cs
@@ -5,10 +5,10 @@ namespace ExchangeApp.BL.Models.Currency;
 public record CurrencyDetailModel : ModelBase
 {
     public required string Code { get; set; }
-    public required float Quantity { get; set; }
+    public required decimal Quantity { get; set; }
     public required string PhotoUrl { get; set; }
-    public required float AverageCourseRate { get; set; }
-    public float? BuyRate { get; set; }
-    public float? SellRate { get; set; }
+    public required decimal AverageCourseRate { get; set; }
+    public decimal? BuyRate { get; set; }
+    public decimal? SellRate { get; set; }
     public required CurrencyState Status { get; set; }
 }
