+++ ExchangeApp.BL/Models/Currency/CurrencyNewTransactionModel.cs
@@ -3,7 +3,7 @@
 public record CurrencyNewTransactionModel : ModelBase
 {
     public required string Code { get; set; }
-    public required float Quantity { get; set; }
-    public required float BuyRate { get; set; }
-    public required float SellRate { get; set; }
+    public required decimal Quantity { get; set; }
+    public required decimal BuyRate { get; set; }
+    public required decimal SellRate { get; set; }
 }
