+++ ExchangeApp.BL/Models/Donation/DonationDetailModel.cs
@@ -8,8 +8,8 @@ public record DonationDetailModel : ModelBase
 {
     public int Id { get; set; }
     public required DateTime Time { get; set; }
-    public required float CourseRate { get; set; }
-    public required float Quantity { get; set; }
+    public required decimal CourseRate { get; set; }
+    public required decimal Quantity { get; set; }
     public required DonationType Type { get; set; }
     public required string Note { get; set; }
     public bool IsCanceled { get; set; }
