+++ ExchangeApp.DAL/Entities/CurrencyEntity.cs
@@ -8,10 +8,10 @@ public record CurrencyEntity : IEntity
 {
     [Key]
     public required string Code { get; set; }
-    public float Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public required string PhotoUrl { get; set; }
-    public float AverageCourseRate { get; set; } = 1;
-    public float? BuyRate { get; set; }
-    public float? SellRate { get; set; }
+    public decimal AverageCourseRate { get; set; } = 1;
+    public decimal? BuyRate { get; set; }
+    public decimal? SellRate { get; set; }
     public CurrencyState Status { get; set; } = CurrencyState.NotInUse;
 }
