+++ ExchangeApp.DAL/Entities/DonationEntity.cs
@@ -7,8 +7,8 @@ public record DonationEntity : IEntity
 {
     public required int Id { get; set; }
     public required DateTime Time { get; set; }
-    public required float CourseRate { get; set; }
-    public required float Quantity { get; set; }
+    public required decimal CourseRate { get; set; }
+    public required decimal Quantity { get; set; }
     public required DonationType Type { get; set; }
     public required string Note { get; set; }
     public bool IsCanceled { get; set; } = false;
